[change][Fix][Refactor] Add delete member feature. Fix no-existed plcace reading undefine exception. Refactor delete popup functions
